backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "magi"
version = "0.1.0"
description = "一个轻量级、可本地部署的AI Agent框架"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Magi Contributors"}
]
keywords = ["ai", "agent", "llm", "autonomous"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    # 核心框架
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",

    # 异步支持
    "asyncio>=3.4.3",

    # 配置管理
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",

    # 日志
    "structlog>=24.1.0",

    # 数据库
    "aiosqlite>=0.19.0",

    # 消息总线（可选Redis）
    "redis>=5.0.1",

    # 向量数据库
    "chromadb>=0.4.22",

    # 图数据库
    "networkx>=3.2.1",

    # LLM SDK
    "openai>=1.10.0",
    "anthropic>=0.18.0",

    # WebSocket
    "python-socketio>=5.11.0",
    "aiohttp>=3.9.0",

    # 认证
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",

    # 工具
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/your-org/magi"
Documentation = "https://magi.readthedocs.io"
Repository = "https://github.com/your-org/magi"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
